changed python version to 3.8 from 3.9, removed flake8 as linter for black and corrected unittest run script
