--- Test/deploy.sh
@@ -1,4 +1,0 @@
-pytest testrunner.py || echo "Tests failed, aborting" && exit 1
-
-clear
-python3 fizzbuzz.py
